backend/pyproject.toml
[tool.poetry]
name = "bridgeapp"
version = "0.1.dev0"
description = "A lightweight web application for playing contract bridge"
authors = ["Jaakko Moisio <jaakko@moisio.fi>"]

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^25.0"
more-itertools = "^9.1"
pydantic = "^1.10"
orjson = "^3.8"
fastapi = "^0.92"
gunicorn = "^20.1"
SQLAlchemy = "^2.0"
SQLAlchemy-Utils = "^0.40"
psycopg2-binary = "^2.9"
passlib = "^1.7.4"
fastpbkdf2 = "^0.2"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
elasticsearch-dsl = "^7.4"
click-log = "^0.4"
click = "^8.0"
hrefs = "^0.8"
asyncpg = "^0.27"

[tool.poetry.dev-dependencies]
pytest = "^7.2"
pytest-asyncio = "^0.20"
black = "^23.1"
pylint = "^2.16"
sphinx = "^6.1"
sphinx-rtd-theme = "^1.2"
coverage = "^7.2"
aiosqlite = "^0.18"
httpx = "^0.23"

[tool.poetry.scripts]
bridgeapp = "bridgeapp.__main__:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
